requirements.txt
django==5.1.5
djangorestframework==3.15.2
firebase-admin==6.6.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
requests==2.32.3
environs
django-cors-headers
python-decouple
pyrebase4
whitenoise
drf-yasg
djangorestframework-simplejwt
eth-utils
boto3
django-storages
dnspython
dj-database-url
psycopg2-binary
supabase